Force everything into the "local" timezone of the conference.
Pretalx was returning correctly constructed timezones that
differed in their UTC offsets. Those offsets were getting lost
when rendered out into the data files and thus all the displayed
times were, to use a technical term, "wonky". This forces all
the times to be in the "local" timezone of the conference thus
ensuring consistency.